refactor: wrap `os_info` package with our own `client_os` interface
`os_info` is a 3rd party package meaning I don't have full control over it, so
I can't e.g. impl traits for it.

Besides, I will need to group OSs in the future e.g. ubuntu or arch should be the same
as linux, and for that I have to have my own interface.