feat(mapping): accurately parse `from` block-level field
Introduce an option to add bash-like `$` variables into the config.

For now, add $TARGET which will be replaced later with the current target from the field; e.g.
if `target` is defined as `linux` and user is currently on linux-like system,
then replace the $TARGET with `linux`, therefore making it possible to customize
`from` field in case standard "files/$TARGET" is not how user wants to organize
dotfiles.